Refactor to move ApolloClient to main.js
