+++ client/src/App.jsx
@@ -1,12 +1,5 @@
 import "./App.css";
 import { Link, Outlet } from "react-router-dom";
-import {
-  ApolloClient,
-  ApolloProvider,
-  InMemoryCache,
-  createHttpLink,
-} from "@apollo/client";
-import { setContext } from "@apollo/client/link/context";
 
 import NavLink from "./components/NavLink.jsx";
 import Navbar from "react-bootstrap/Navbar";
@@ -24,77 +17,56 @@ import User from "./utils/user.js";
 import { useThemeContext } from "./components/ThemeContext.jsx";
 import { useUserContext } from "./components/UserContext.jsx";
 
-const httpLink = createHttpLink({
-  uri: "/graphql",
-});
-
-const authLink = setContext((_, { headers }) => {
-  const token = localStorage.getItem("auth_token");
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : "",
-    },
-  };
-});
-
-const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
-
 function App() {
   const { theme, setTheme } = useThemeContext();
   const { user, setUser } = useUserContext();
 
   return (
     <div className={theme}>
-      <ApolloProvider client={client}>
-        <Navbar expand="lg" id="header">
-          <Container>
-            <Link
-              to={"/"}
-              className="d-flex align-items-center text-decoration-none"
-            >
-              <Image src={theme === "dark" ? darkLogo : Logo} fluid></Image>
-              <h1>Money Type</h1>
-            </Link>
-            <nav>
-              <NavLink to={"/leaderboard"}>Leaderboard</NavLink>
-              {user._id && User.isLoggedIn() ? (
-                <button
-                  as="input"
-                  type="button"
-                  className="navbtn"
-                  onClick={() => {
-                    User.logout();
-                    setUser({});
-                  }}
-                >
-                  Logout
-                </button>
-              ) : (
-                <NavLink to={`/signUp`}>Sign Up</NavLink>
-              )}
-            </nav>
-          </Container>
-        </Navbar>
-        <div className="gameCard">
-          <Outlet />
-        </div>
+      <Navbar expand="lg" id="header">
         <Container>
-          <div className="toggle d-flex justify-content-end">
-            <button
-              onClick={() => {
-                setTheme(theme === "dark" ? "light" : "dark");
-              }}
-              className="theme-switcher"
-            >
-              <Image src={theme === "dark" ? moonSvg : sunSvg} fluid></Image>
-            </button>
-          </div>
+          <Link
+            to={"/"}
+            className="d-flex align-items-center text-decoration-none"
+          >
+            <Image src={theme === "dark" ? darkLogo : Logo} fluid></Image>
+            <h1>Money Type</h1>
+          </Link>
+          <nav>
+            <NavLink to={"/leaderboard"}>Leaderboard</NavLink>
+            {user._id && User.isLoggedIn() ? (
+              <button
+                as="input"
+                type="button"
+                className="navbtn"
+                onClick={() => {
+                  User.logout();
+                  setUser({});
+                }}
+              >
+                Logout
+              </button>
+            ) : (
+              <NavLink to={`/signUp`}>Sign Up</NavLink>
+            )}
+          </nav>
         </Container>
-      </ApolloProvider>
+      </Navbar>
+      <div className="gameCard">
+        <Outlet />
+      </div>
+      <Container>
+        <div className="toggle d-flex justify-content-end">
+          <button
+            onClick={() => {
+              setTheme(theme === "dark" ? "light" : "dark");
+            }}
+            className="theme-switcher"
+          >
+            <Image src={theme === "dark" ? moonSvg : sunSvg} fluid></Image>
+          </button>
+        </div>
+      </Container>
     </div>
   );
 }
